+++ packages/gestalt/src/Checkbox.js
@@ -15,6 +15,10 @@ type Props = {|
   indeterminate?: boolean,
   name?: string,
   onChange: ({ event: SyntheticInputEvent<>, checked: boolean }) => void,
+  onClick?: ({
+    event: SyntheticInputEvent<HTMLInputElement>,
+    checked: boolean,
+  }) => void,
   size?: 'sm' | 'md',
 |};
 
@@ -31,6 +35,7 @@ export default class Checkbox extends React.Component<Props, State> {
     indeterminate: PropTypes.bool,
     name: PropTypes.string,
     onChange: PropTypes.func.isRequired,
+    onClick: PropTypes.func,
     size: PropTypes.oneOf(['sm', 'md']),
   };
 
@@ -67,6 +72,14 @@ export default class Checkbox extends React.Component<Props, State> {
   handleChange = (event: SyntheticInputEvent<>) => {
     const { checked } = event.target;
     this.props.onChange({ event, checked });
+  };
+
+  handleClick = (event: SyntheticInputEvent<HTMLInputElement>) => {
+    const { onClick } = this.props;
+    if (onClick) {
+      const { checked } = event.currentTarget;
+      onClick({ event, checked });
+    }
   };
 
   handleBlur = () => this.setState({ focused: false });
@@ -108,6 +121,7 @@ export default class Checkbox extends React.Component<Props, State> {
           name={name}
           onBlur={this.handleBlur}
           onChange={this.handleChange}
+          onClick={this.handleClick}
           onFocus={this.handleFocus}
           ref={el => {
             this.input = el;