+++ CHANGELOG.md
@@ -11,6 +11,7 @@
 - Icon: Add compose icon (#358)
 - SearchField: Add `autoComplete` prop for parity with `TextField` (#363)
 - Tabs: Add optional wrap prop (#361)
+- Checkbox: Add optional onClick prop (#364)
 
 ### Patch